==> notebooks/3_one_hot_encoding_cleaning.ipynb <==
# 3. One-hot encoding for the genres and cleaning the data.

## Importing the data and cleaning the data.

import pandas as pd

df = pd.read_csv('../data/5_Bru_spotify_million_tracks.csv')

df.head()


# Remove '[' and ']' and single quotes from the 'genres' column
df['genres'] = df['genres'].apply(lambda x: str(x).replace('[', '').replace(']', '').replace("'", '') if pd.notnull(x) else '')
df.drop(columns=['album_cover'], inplace=True)
df.head()

# Extract the first 4 characters (year) and convert to integer
df['release_date'] = df['release_date'].str[:4].astype(int)

df.head()


## One-hot encoding the genres column.

df['genres'] = df['genres'].str.split(',')

#Perform one-hot encoding
genres_one_hot = df['genres'].explode().str.strip().str.get_dummies().groupby(level=0).sum()

#Merge the one-hot encoded columns back into the original DataFrame
df = pd.concat([df, genres_one_hot], axis=1)

#Drop the original 'genres' column if no longer needed
df = df.drop(columns=['genres'])

df.head()


## Cleaning the data.

df.shape

# Check if a specific genre is encoded as 1 for any song
rock_songs = df[df['"childrens music"'] == 1]
print("Songs with the genre 'rock':")
rock_songs

# Convert 'explicit'columns to 1s and 0s
df['explicit'] = df['explicit'].astype(int)
df.head()

df.shape

df.head()

# Remove double quotes from column names
df.columns = df.columns.str.replace('"', '')
df.head()

df.to_csv('../data/6_Bru_one_hot_encoded_is_hot_cleaned.csv', index=False)

# Drop the columns that are not numerical.
df.drop(columns=['original_title', 'original_artist', 'spotify_title', 'spotify_artist', 'album'], inplace=True)

df.head()

df.to_csv('../data/7_Bru_one_hot_encoded_is_hot_cleaned_numerical.csv', index=False)
